UVC: Add bulk mode support on uvc
Add a config in
/sys/class/plcm_usb/plcm0/f_webcam/webcam_bulkmode.

If the bulkmode is set to 1, we will use the BULK EP
to transfer the video stream. So modified some video
streaming descriptors. Unlike the iso endpoints, the
bulk ep only has the alternate 0 and it wil not receive
the set_alt 1 request.

In bulk mode, the wMaxPayloadTransferSize is set to
16 * 1024 by the original uvc author. It works, but
we still to figure out the reason.

Change-Id: Id34519167fd497939e10e4053374c467ddea6fcb