UDM: Add a USB Device Manager Daemon
It depends on the libudev and libdbus-1.

It listens to the uevent coming from
"/sys/devices/virtual/plcm_usb/plcm0", and then send the usb_state
signal to the dbus session bus. So that any other dbus clients can
recevie it.

Also it handles some method calls from the dbus. Including
GetCurrentUsbState
GetDefaultUsbFunctions
SetCurrentUsbFunctions
StartUsbDevice
StopUsbDevice

It uses the System Bus, so we need to add a policy file to gain
the access. The policy file should be copied to
/etc/dbus-1/system.d/

Added a client sample in test/.

Change-Id: I2e10ff2b09bf7af40d67e9af9f910bd8844f308e