SAMPLE: Send out signals while handling the VC requests
In handling of the video control requests from the host, such as
changing Brightness, Contrast, the uvc-gadget daemon will post
dbus signals to app side. Then the app will start a method call
to send the value back.

Also provided a python script which demonstrates how to handle
the signals and call the method call.

Change-Id: Ib26251be9d6cc227ec39d35b907e49b533cd5d59